Move `apply_patch()` to reduce chance for circular imports
This could happen when core components like the `ConfigManager` are
patched.

This changeset moves the set of enabled patches into a dedicated
leaf-module. And move the `apply_patch()` helper into the `patches`
module directly.

Imports are adjusted and a legacy import is kept in place. The
documentation on patching is updated accordingly.
