+++ datalad_next/patches/annexrepo.py
@@ -29,7 +29,7 @@ from datalad_next.utils import (
     get_specialremote_credential_properties,
     needs_specialremote_credential_envpatch,
 )
-from datalad_next.utils.patch import apply_patch
+from . import apply_patch
 
 
 # reuse logger from -core, despite the unconventional name
