+++ datalad_next/patches/enabled.py
@@ -1,0 +1,13 @@
+from . import (
+    commanderror,
+    common_cfg,
+    annexrepo,
+    configuration,
+    create_sibling_ghlike,
+    interface_utils,
+    push_to_export_remote,
+    push_optimize,
+    siblings,
+    test_keyring,
+    customremotes_main,
+)
