+++ datalad_next/patches/siblings.py
@@ -17,7 +17,7 @@ from datalad.support.exceptions import (
     AccessFailedError,
     CapturedException,
 )
-from datalad_next.utils.patch import apply_patch
+from . import apply_patch
 
 
 # use same logger as -core
